DEMO: Pull errors into a separate crate
The only thing used in `primitives` and `hashes` from the `hex` crate is
the error types (actually just the `HexToArrayError` and inner types).

In an effort to stabalise more quickly pull the error types out into a
separate crate. Set the version to `1.0.0-alpha.0` and release it
already.
