+++ errors/CHANGELOG.md
@@ -1,0 +1,4 @@
+# 1.0.0-alpha.0 - 2025-02-21
+
+Initial release. This code was pulled out of `hex-conservative` and is
+intended to stabalize immediately.
